backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
feedparser>=6.0.11
pydantic>=2.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
scikit-learn>=1.3.0
